Made switching to detailed screen possible
